refactor: move nc to its own file to avoid an import cycle
