+++ client/src/components/AutocompleteInput/nc.js
@@ -1,0 +1,3 @@
+import { namespaceClassName } from "../../utils/props";
+
+export const nc = namespaceClassName("c-wagtailautocomplete");
